+++ crates/ferroid/src/futures/mod.rs
@@ -1,7 +1,7 @@
 mod snowflake;
-#[cfg(test)]
-mod tests;
 mod ulid;
+mod sleep_provider;
 
 pub use snowflake::*;
 pub use ulid::*;
+pub use sleep_provider::*;
